refactor: FSD 의존성과 공개 API 경계 정리 (#161)
