+++ src/features/chat/index.ts
@@ -1,3 +1,10 @@
 export * from "./types";
+export {
+  addMessageToCache,
+  hasMultipleDates,
+  isSameUserContinuous,
+  removeMatchingTempMessage,
+  runGarbageCollection,
+} from "./lib/messageUtils";
 export { default as ChatHistory } from "./ui/ChatHistory";
 export { default as MessageField } from "./ui/MessageField";
