+++ src/features/presence/types.ts
@@ -13,3 +13,31 @@ export interface PresenceParticipant {
   /** 발표자의 마이크가 실제로 활성화되어 있는지 여부 */
   audioEnabled?: boolean;
 }
+
+/**
+ * Supabase Presence track payload 구조를 정의한다.
+ */
+export interface PresenceTrackPayload {
+  userId: string;
+  nickname: string;
+  villageId: VillageId;
+  joinedAt: string;
+  username?: string;
+  isSpeaker?: boolean;
+  /** 음성 채널에 연결되어 있는지 여부 */
+  voiceConnected?: boolean;
+  /** 발표자의 마이크가 실제로 활성화되어 있는지 여부 */
+  audioEnabled?: boolean;
+}
+
+/**
+ * Supabase presenceState() raw item 구조를 정의한다.
+ */
+export interface PresenceStateItem extends Partial<Omit<PresenceTrackPayload, "villageId">> {
+  presence_ref: string;
+  villageId?: string | null;
+  user_id?: string;
+  user_nickname?: string;
+  joined_at?: string;
+  online_at?: string;
+}
