--- src/shared/lib/chat/index.ts
@@ -1,7 +1,0 @@
-export {
-  hasMultipleDates,
-  isSameUserContinuous,
-  removeMatchingTempMessage,
-  addMessageToCache,
-  runGarbageCollection,
-} from "./messageUtils";
