+++ src/shared/lib/index.ts
@@ -1,10 +1,3 @@
-export {
-  hasMultipleDates,
-  isSameUserContinuous,
-  removeMatchingTempMessage,
-  addMessageToCache,
-  runGarbageCollection,
-} from "./chat";
 export { toDate, isSameDay, formatTime, formatDate, formatJoinedTime } from "./datetime";
 export { getChatChannelName } from "./realtime/getChatChannelName";
 export { getChatRoomId } from "./realtime/getChatRoomId";
