+++ src/widgets/chatPanel/model/useChatPanel.ts
@@ -1,17 +1,17 @@
 "use client";
 
 import { useSupabase } from "@/app/providers/SupabaseProvider";
-import { Message, MessagesPage } from "@/features/chat";
+import {
+  Message,
+  MessagesPage,
+  addMessageToCache,
+  removeMatchingTempMessage,
+  runGarbageCollection,
+} from "@/features/chat";
 import { useMessagesQuery } from "@/features/chat/hooks/useMessagesQuery";
 import { useMovementStore } from "@/features/movement/model/useMovementStore";
 import { CHAT_GC_CONFIG, CHAT_TABLE_NAME } from "@/shared/config";
-import {
-  addMessageToCache,
-  getChatChannelName,
-  getChatRoomId,
-  removeMatchingTempMessage,
-  runGarbageCollection,
-} from "@/shared/lib";
+import { getChatChannelName, getChatRoomId } from "@/shared/lib";
 import { useChatVisibilityActions, useUserStore, useVisiblePageIndices } from "@/shared/store";
 import { InfiniteData, useQueryClient } from "@tanstack/react-query";
